\name{kdml}
\alias{kdml}
\alias{kdml-package}
\docType{package}
\title{Kernel Metric Learning for Mixed-type Data}
\description{

This package provides nonparametric kernel methods for calculating pairwise distances between mixed-type observations. Returning a numeric pairwise distance matrix for usage in clustering or classification applications is the intended purpose. Descriptions of the implementation of these methods can be found in Ghashti, J.S. (2024) and Ghashti, J.S. and Thompson, J.R.J. (2023).

}

\details{

This package contains two functions for pairwise distance calculations of mixed-type data based on two different methods. Kernel methods also require variable-specific bandwidths, with two additional functions for the bandwidth specification methods.

}

\author{
  Jesse S. Ghashti <jesse.ghashti@ubc.ca>

  Maintainer: John R.J. Thompson <john.thompson@ubc.ca>

  We would like to acknowledge funding support from the University of British Columbia Aspire Fund (UBC:www.ok.ubc.ca/). We also acknowledge the support of the Natural Sciences and Engineering Research Council of Canada (NSERC). 

}
\references{

     Ghashti, J.S. (2024), \emph{Similarity Maximization and Shrinkage Approach in Kernel Metric Learning for Clustering Mixed-type Data (T)}, University of British Columbia. <https://dx.doi.org/10.14288/1.044397>

     Ghashti, J.S. and J.R.J Thompson (2023), \dQuote{Mixed-type Distance Shrinkage and Selection for Clustering via Kernel Metric Learning.} arXiv preprint arXiv:2306.01890. <
https://doi.org/10.48550/arXiv.2306.01890>


}
\keyword{package}
